+++ packages/profile-sync-controller/src/controllers/user-storage/UserStorageController.ts
@@ -283,7 +283,6 @@ export default class UserStorageController extends BaseController<
   // This is replaced with the actual value in the constructor
   // We will remove this once the feature will be released
   readonly #env = {
-    isAccountSyncingEnabled: false,
     isNetworkSyncingEnabled: false,
   };
 
@@ -342,7 +341,6 @@ export default class UserStorageController extends BaseController<
     state?: UserStorageControllerState;
     config?: ControllerConfig;
     env?: {
-      isAccountSyncingEnabled?: boolean;
       isNetworkSyncingEnabled?: boolean;
     };
     nativeScryptCrypto?: NativeScrypt;
@@ -354,7 +352,6 @@ export default class UserStorageController extends BaseController<
       state: { ...defaultState, ...state },
     });
 
-    this.#env.isAccountSyncingEnabled = Boolean(env?.isAccountSyncingEnabled);
     this.#env.isNetworkSyncingEnabled = Boolean(env?.isNetworkSyncingEnabled);
     this.#config = config;
 
@@ -391,15 +388,10 @@ export default class UserStorageController extends BaseController<
     this.#nativeScryptCrypto = nativeScryptCrypto;
 
     // Account Syncing
-    if (this.#env.isAccountSyncingEnabled) {
-      setupAccountSyncingSubscriptions(
-        { isAccountSyncingEnabled: true },
-        {
-          getUserStorageControllerInstance: () => this,
-          getMessenger: () => this.messagingSystem,
-        },
-      );
-    }
+    setupAccountSyncingSubscriptions({
+      getUserStorageControllerInstance: () => this,
+      getMessenger: () => this.messagingSystem,
+    });
 
     // Network Syncing
     if (this.#env.isNetworkSyncingEnabled) {
@@ -718,7 +710,6 @@ export default class UserStorageController extends BaseController<
 
     await syncInternalAccountsWithUserStorage(
       {
-        isAccountSyncingEnabled: this.#env.isAccountSyncingEnabled,
         maxNumberOfAccountsToAdd:
           this.#config?.accountSyncing?.maxNumberOfAccountsToAdd,
         onAccountAdded: () =>
@@ -747,14 +738,10 @@ export default class UserStorageController extends BaseController<
   async saveInternalAccountToUserStorage(
     internalAccount: InternalAccount,
   ): Promise<void> {
-    await saveInternalAccountToUserStorage(
-      internalAccount,
-      { isAccountSyncingEnabled: this.#env.isAccountSyncingEnabled },
-      {
-        getMessenger: () => this.messagingSystem,
-        getUserStorageControllerInstance: () => this,
-      },
-    );
+    await saveInternalAccountToUserStorage(internalAccount, {
+      getMessenger: () => this.messagingSystem,
+      getUserStorageControllerInstance: () => this,
+    });
   }
 
   async syncNetworks() {